requirements.txt

fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
httpx
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
